Split NonEmptyEnumerable's nested types into their own files
Move the nested Enumerator struct to NonEmptyEnumerable.Enumerator.cs
(partial class) and the DebugView to NonEmptyEnumerableDebugView.cs, so
NonEmptyEnumerable.cs stays focused on the core collection surface.
